dev-requirements.txt
pytest >=4.6
coverage >= 5.2.1
pytest-cov >= 2.10.0
bandit >= 1.6.2
flake8 >= 3.8.3
isort >= 5.4.1
black >= 22.1
autopep8 >= 1.5.5
pyspark==3.3.2
delta-spark==2.3.0
Sphinx==4.5.0
sphinxcontrib-confluencebuilder==2.0.0
sphinx-rtd-theme==1.2.0